requirements.txt
asgiref==3.5.2
bootstrap4==0.1.0
Django==4.1.3
django-crispy-forms==1.14.0
django-environ==0.9.0
django-extensions==3.2.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
mysqlclient==2.1.1
python-dotenv==1.0.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
